+++ incore_auth/app.py
@@ -2,11 +2,15 @@ import logging
 import json
 import os
 import time
+import threading
 import urllib.request
 
 import IP2Location
 import geohash2
 import influxdb_client
+import pymongo
+
+from cachetools import cached, TTLCache
 
 from flask import Flask, request, Response, make_response, json
 from jose import jwt
@@ -25,6 +29,10 @@ app.config.from_mapping(config)
 geoserver = {}
 geoserver_delta = 2
 
+cache_size = 1024
+# timeout in seconds, in this case 30 minutes
+cache_timeout = 30*60
+
 # setup database for geolocation
 try:
     geolocation = IP2Location.IP2Location(CONTRIBUTION_DB_NAME)
@@ -38,6 +46,87 @@ if __name__ != '__main__':
     gunicorn_logger = logging.getLogger('gunicorn.error')
     app.logger.handlers = gunicorn_logger.handlers
     app.logger.setLevel(gunicorn_logger.level)
+
+
+def cache_key(request_info):
+    """Return username from request_info to be used as cache key"""
+    return request_info["username"]
+
+
+def update_services_thread(request_info):
+    """When a user does any action, it will check to update the groups in mongo, as well as make sure
+    the user has access to datawolf. The function is cached to make sure only every 30 minutes we do
+    the checks, since this can be expensive."""
+
+    # get information from request
+    username = request_info["username"]
+    if not username:
+        return username
+    groups = request_info['groups']
+
+    # call datawolf to add user
+    datawolf_url = config["datawolf_url"]
+    if datawolf_url:
+        query = urllib.parse.urlencode({
+            "firstname": request_info["firstname"],
+            "lastname": request_info["lastname"],
+            "email": username
+        })
+        datawolf_url = "%s/persons?%s" % (datawolf_url.rstrip("/"), query)
+        req = urllib.request.Request(datawolf_url, method='POST')
+        response = urllib.request.urlopen(req)
+        if response.code == 200:
+            app.logger.info(f"Added user to datawolf {username}")
+        elif response.code == 204:
+            app.logger.debug(f"User already exists in datawolf {username}")
+        else:
+            app.logger.info(f"Did not add user to datawolf {username}")
+
+    # update database with user quota
+    mongo_client = config["mongo_client"]
+    if mongo_client:
+        mongo_user = mongo_client["spacedb"]["UserGroups"].find_one({"username": username})
+        if not mongo_user:
+            # INSERT
+            mongo_client["spacedb"]["UserGroups"].insert_one({
+                "username": username,
+                "className": "edu.illinois.ncsa.incore.common.models.UserGroups",
+                "groups": groups
+            })
+            app.logger.info(f"Inserted groups document for {username}")
+        elif set(groups) != set(mongo_user["groups"]):
+            # UPDATE
+            mongo_client["spacedb"]["UserGroups"].update_one(
+                {"username": username}, {"$set": {"groups": groups}}
+            )
+            app.logger.info(f"Synced groups for {username} - {groups}")
+        else:
+            # NOTHING
+            app.logger.debug(f"No sync needed for {username}")
+
+        mongo_space = mongo_client["spacedb"]["Space"].find_one({"metadata.name": username})
+        if not mongo_space:
+            mongo_client["spacedb"]["Space"].insert_one({
+                "className": "edu.illinois.ncsa.incore.common.models.Space",
+                "metadata": {
+                    "className": "edu.illinois.ncsa.incore.common.models.SpaceMetadata",
+                    "name": username
+                },
+                "privileges": {
+                    "className": "edu.illinois.ncsa.incore.common.auth.Privileges",
+                    "userPrivileges": {
+                        username: "ADMIN"
+                    }
+                },
+                "members": [
+                ]
+            })
+            app.logger.info(f"Inserted space document for {username}")
+
+
+@cached(cache=TTLCache(maxsize=cache_size, ttl=cache_timeout), key=cache_key)
+def update_services(request_info):
+    threading.Thread(target=update_services_thread, args=(request_info,), daemon=True).start()
 
 
 def record_request(request_info):
@@ -173,6 +262,11 @@ def request_userinfo(request_info):
         request_info['error'] = 'JWT Error: invalid token'
         return
 
+    # get name of user
+    request_info["firstname"] = access_token["given_name"]
+    request_info["lastname"] = access_token["family_name"]
+    request_info["fullname"] = access_token["name"]
+
     # retrieve the groups the user belongs to from access token
     request_info['username'] = access_token["preferred_username"]
     request_info['groups'] = access_token.get("groups", [])
@@ -200,6 +294,8 @@ def request_resource(request_info):
             request_info['resource'] = pieces[1]
             if request_info['resource'] == "doc" and len(pieces) > 2:
                 request_info['fields']['manual'] = pieces[2]
+            if request_info['resource'] == "playbook" and len(pieces) > 2:
+                request_info['fields']['playbook'] = pieces[2]
             if request_info['resource'] == "data" and len(pieces) > 4 and uri.endswith('blob'):
                 request_info['fields']['dataset'] = pieces[4]
             if request_info['resource'] == "dfr3" and len(pieces) > 4:
@@ -238,6 +334,9 @@ def verify_token():
     # dict to hold all information
     request_info = {
         "username": "",
+        "firstname": "",
+        "lastname": "",
+        "fullname": "",
         "method": request.method,
         "url": request.path,
         "resource": "",
@@ -252,6 +351,9 @@ def verify_token():
     # get info requested
     request_resource(request_info)
     request_userinfo(request_info)
+
+    # update backend services
+    update_services(request_info)
 
     # record request
     record_request(request_info)
@@ -339,6 +441,17 @@ def setup():
         config['audience'] = keycloak_audience
     else:
         config['audience'] = None
+
+    # store datawolf url
+    config["datawolf_url"] = os.environ.get('DATAWOLF_URL', None)
+
+    # setup mongodb
+    mongodb_uri = os.environ.get('MONGODB_URI', None)
+    if mongodb_uri:
+        mongo_client = pymongo.MongoClient(mongodb_uri)
+        config["mongo_client"] = mongo_client
+    else:
+        config["mongo_client"] = None
 
     # setup influxdb
     try: